third_party/WebKit/LayoutTests/svg/dom/SVGTransformList-empty-list-consolidation.html

<!DOCTYPE html>
<title>SVGTransformList.consolidate() on an empty transform list</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {
  let element = document.createElementNS("http://www.w3.org/2000/svg", "pattern");
  let transformList = element.patternTransform.baseVal;
  assert_equals(transformList.length, 0);
  assert_equals(transformList.consolidate(), null);
});
</script>
